:recycle: Chore: Move debug to it's own namespace away from anvil
