+++ vm/api/src/handlers/index.ts
@@ -2,5 +2,6 @@ export type * from './AccountHandler.js'
 export type * from './CallHandler.js'
 export type * from './ContractHandler.js'
 export type * from './ScriptHandler.js'
-export type * from './EthHandler.js'
 export type * from './AnvilHandler.js'
+export type * from './DebugHandler.js'
+export type * from './EthHandler.js'
