+++ vm/api/src/params/index.ts
@@ -5,3 +5,4 @@ export type * from './ContractParams.js'
 export type * from './ScriptParams.js'
 export type * from './EthParams.js'
 export type * from './AnvilParams.js'
+export type * from './DebugParams.js'
