+++ vm/api/src/procedure/index.ts
@@ -5,3 +5,4 @@ export type * from './ScriptJsonRpcProcedure.js'
 export type * from './EthProcedure.js'
 export type * from './EthProcedure.js'
 export type * from './AnvilProcedure.js'
+export type * from './DebugProcedure.js'
