+++ vm/api/src/requests/index.ts
@@ -6,3 +6,4 @@ export * from './ScriptJsonRpcRequest.js'
 export * from './TevmJsonRpcRequest.js'
 export * from './EthJsonRpcRequest.js'
 export * from './AnvilJsonRpcRequest.js'
+export * from './DebugJsonRpcRequest.js'
