+++ vm/api/src/responses/index.ts
@@ -5,3 +5,4 @@ export * from './JsonRpcResponse.js'
 export * from './ScriptJsonRpcResponse.js'
 export * from './EthJsonRpcResponse.js'
 export * from './AnvilJsonRpcResponse.js'
+export * from './DebugJsonRpcResponse.js'
