+++ vm/api/src/result/index.ts
@@ -4,3 +4,4 @@ export type * from './ContractResult.js'
 export type * from './ScriptResult.js'
 export type * from './EthResult.js'
 export type * from './AnvilResult.js'
+export type * from './DebugResult.js'
