+++ vm/procedures/src/handlers/debug/traceCall.js
@@ -10,78 +10,76 @@ import { bytesToHex, hexToBytes, numberToHex } from 'viem'
  */
 export const traceCallHandler =
 	({ vm: _vm }) =>
-	async (params) => {
-		// const block = await getBlockByOption(blockOpt, this.chain)
-		// const parentBlock = await chain.getBlock(block.header.parentHash)
+		async (params) => {
+			const vm = await _vm.shallowCopy()
 
-		const vm = await _vm.shallowCopy()
-		// await vm.stateManager.setStateRoot(parentBlock.header.stateRoot)
-		const {
-			account,
-			to,
-			gas: gasLimit,
-			gasPrice,
-			value,
-			data,
-		} = params.transaction
+			const {
+				account,
+				to,
+				gas: gasLimit,
+				gasPrice,
+				value,
+				data,
+			} = params.transaction
 
-		const trace = {
-			gas: 0n,
-			/**
-			 * @type {import('viem').Hex}
-			 */
-			returnValue: '0x0',
-			failed: false,
-			/**
-			 * @type {Array<import('@tevm/api').DebugTraceCallResult['structLogs'][number]>}
-			 */
-			structLogs: [],
-		}
-		vm.evm.events?.on('step', async (step, next) => {
-			trace.structLogs.push({
-				pc: step.pc,
-				op: step.opcode.name,
-				gasCost: BigInt(step.opcode.fee) + (step.opcode.dynamicFee ?? 0n),
-				gas: step.gasLeft,
-				depth: step.depth,
-				// TODO make disableStack disableStack
-				// ...(params.tracerConfig?.disableStack !== true ? {stack: step.stack.map(code => numberToHex(code))} : {}),
-				stack: step.stack.map((code) => numberToHex(code)),
-			})
-			next?.()
-		})
-
-		vm.evm.events?.on('afterMessage', (data, next) => {
-			if (
-				data.execResult.exceptionError !== undefined &&
-				trace.structLogs.length > 0
-			) {
-				// Mark last opcode trace as error if exception occurs
-				const nextLog = trace.structLogs[trace.structLogs.length - 1]
-				if (!nextLog) {
-					throw new Error('No structLogs to mark as error')
-				}
-				// nextLog.error = true
+			const trace = {
+				gas: 0n,
+				/**
+				 * @type {import('viem').Hex}
+				 */
+				returnValue: '0x0',
+				failed: false,
+				/**
+				 * @type {Array<import('@tevm/api').DebugTraceCallResult['structLogs'][number]>}
+				 */
+				structLogs: [],
 			}
-			next?.()
-		})
+			vm.evm.events?.on('step', async (step, next) => {
+				trace.structLogs.push({
+					pc: step.pc,
+					op: step.opcode.name,
+					gasCost: BigInt(step.opcode.fee) + (step.opcode.dynamicFee ?? 0n),
+					gas: step.gasLeft,
+					depth: step.depth,
+					stack: step.stack.map((code) => numberToHex(code)),
+				})
+				next?.()
+			})
 
-		const res = await vm.evm.runCall({
-			...(account !== undefined
-				? {
+			vm.evm.events?.on('afterMessage', (data, next) => {
+				if (
+					data.execResult.exceptionError !== undefined &&
+					trace.structLogs.length > 0
+				) {
+					// Mark last opcode trace as error if exception occurs
+					const nextLog = trace.structLogs[trace.structLogs.length - 1]
+					if (!nextLog) {
+						throw new Error('No structLogs to mark as error')
+					}
+					// TODO fix this type
+					Object.assign(nextLog, {
+						error: data.execResult.exceptionError,
+					})
+				}
+				next?.()
+			})
+
+			const res = await vm.evm.runCall({
+				...(account !== undefined
+					? {
 						caller: Address.fromString(
 							typeof account === 'string' ? account : account.address,
 						),
-				  }
-				: {}),
-			...(data !== undefined ? { data: hexToBytes(data) } : {}),
-			...(to ? { to: Address.fromString(to) } : {}),
-			...(gasPrice ? { gasPrice } : {}),
-			...(gasLimit ? { gasLimit } : {}),
-			...(value ? { value } : {}),
-		})
-		trace.gas = res.execResult.executionGasUsed
-		trace.failed = res.execResult.exceptionError !== undefined
-		trace.returnValue = bytesToHex(res.execResult.returnValue)
-		return trace
-	}
+					}
+					: {}),
+				...(data !== undefined ? { data: hexToBytes(data) } : {}),
+				...(to ? { to: Address.fromString(to) } : {}),
+				...(gasPrice ? { gasPrice } : {}),
+				...(gasLimit ? { gasLimit } : {}),
+				...(value ? { value } : {}),
+			})
+			trace.gas = res.execResult.executionGasUsed
+			trace.failed = res.execResult.exceptionError !== undefined
+			trace.returnValue = bytesToHex(res.execResult.returnValue)
+			return trace
+		}
